requirements.txt
cx_Oracle==8.3.0
python-dotenv==1.0.0
